+++ nautilus_trader/common/factories.pyx
@@ -202,6 +202,7 @@ cdef class OrderFactory:
         Quantity quantity,
         TimeInForce time_in_force = TimeInForce.GTC,
         bint reduce_only = False,
+        bint quote_quantity = False,
         ExecAlgorithmId exec_algorithm_id = None,
         dict exec_algorithm_params = None,
         str tags = None,
@@ -221,6 +222,8 @@ cdef class OrderFactory:
             The orders time in force. Often not applicable for market orders.
         reduce_only : bool, default False
             If the order carries the 'reduce-only' execution instruction.
+        quote_quantity : bool
+            If the order quantity is denominated in the quote currency.
         exec_algorithm_id : ExecAlgorithmId, optional
             The execution algorithm ID for the order.
         exec_algorithm_params : dict[str, Any], optional
@@ -250,6 +253,7 @@ cdef class OrderFactory:
             quantity=quantity,
             time_in_force=time_in_force,
             reduce_only=reduce_only,
+            quote_quantity=quote_quantity,
             init_id=UUID4(),
             ts_init=self._clock.timestamp_ns(),
             contingency_type=ContingencyType.NO_CONTINGENCY,
@@ -271,6 +275,7 @@ cdef class OrderFactory:
         datetime expire_time = None,
         bint post_only = False,
         bint reduce_only = False,
+        bint quote_quantity = False,
         Quantity display_qty = None,
         TriggerType emulation_trigger = TriggerType.NO_TRIGGER,
         InstrumentId trigger_instrument_id = None,
@@ -299,6 +304,8 @@ cdef class OrderFactory:
             If the order will only provide liquidity (make a market).
         reduce_only : bool, default False
             If the order carries the 'reduce-only' execution instruction.
+        quote_quantity : bool
+            If the order quantity is denominated in the quote currency.
         display_qty : Quantity, optional
             The quantity of the order to display on the public book (iceberg).
         emulation_trigger : TriggerType, default ``NO_TRIGGER``
@@ -341,6 +348,7 @@ cdef class OrderFactory:
             expire_time_ns=0 if expire_time is None else dt_to_unix_nanos(expire_time),
             post_only=post_only,
             reduce_only=reduce_only,
+            quote_quantity=quote_quantity,
             display_qty=display_qty,
             emulation_trigger=emulation_trigger,
             trigger_instrument_id=trigger_instrument_id,
@@ -363,6 +371,7 @@ cdef class OrderFactory:
         TimeInForce time_in_force = TimeInForce.GTC,
         datetime expire_time = None,
         bint reduce_only = False,
+        bint quote_quantity = False,
         TriggerType emulation_trigger = TriggerType.NO_TRIGGER,
         InstrumentId trigger_instrument_id = None,
         ExecAlgorithmId exec_algorithm_id = None,
@@ -390,6 +399,8 @@ cdef class OrderFactory:
             The order expiration (for ``GTD`` orders).
         reduce_only : bool, default False
             If the order carries the 'reduce-only' execution instruction.
+        quote_quantity : bool
+            If the order quantity is denominated in the quote currency.
         emulation_trigger : TriggerType, default ``NO_TRIGGER``
             The orders emulation trigger.
         trigger_instrument_id : InstrumentId, optional
@@ -432,6 +443,7 @@ cdef class OrderFactory:
             time_in_force=time_in_force,
             expire_time_ns=0 if expire_time is None else dt_to_unix_nanos(expire_time),
             reduce_only=reduce_only,
+            quote_quantity=quote_quantity,
             emulation_trigger=emulation_trigger,
             trigger_instrument_id=trigger_instrument_id,
             contingency_type=ContingencyType.NO_CONTINGENCY,
@@ -455,6 +467,7 @@ cdef class OrderFactory:
         datetime expire_time = None,
         bint post_only = False,
         bint reduce_only = False,
+        bint quote_quantity = False,
         Quantity display_qty = None,
         TriggerType emulation_trigger = TriggerType.NO_TRIGGER,
         InstrumentId trigger_instrument_id = None,
@@ -487,6 +500,8 @@ cdef class OrderFactory:
             If the order will only provide liquidity (make a market).
         reduce_only : bool, default False
             If the order carries the 'reduce-only' execution instruction.
+        quote_quantity : bool
+            If the order quantity is denominated in the quote currency.
         display_qty : Quantity, optional
             The quantity of the order to display on the public book (iceberg).
         emulation_trigger : TriggerType, default ``NO_TRIGGER``
@@ -535,6 +550,7 @@ cdef class OrderFactory:
             expire_time_ns=0 if expire_time is None else dt_to_unix_nanos(expire_time),
             post_only=post_only,
             reduce_only=reduce_only,
+            quote_quantity=quote_quantity,
             display_qty=display_qty,
             emulation_trigger=emulation_trigger,
             trigger_instrument_id=trigger_instrument_id,
@@ -555,6 +571,7 @@ cdef class OrderFactory:
         TimeInForce time_in_force = TimeInForce.GTC,
         datetime expire_time = None,
         bint reduce_only = False,
+        bint quote_quantity = False,
         Quantity display_qty = None,
         ExecAlgorithmId exec_algorithm_id = None,
         dict exec_algorithm_params = None,
@@ -577,6 +594,8 @@ cdef class OrderFactory:
             The order expiration (for ``GTD`` orders).
         reduce_only : bool, default False
             If the order carries the 'reduce-only' execution instruction.
+        quote_quantity : bool
+            If the order quantity is denominated in the quote currency.
         display_qty : Quantity, optional
             The quantity of the limit order to display on the public book (iceberg).
         exec_algorithm_id : ExecAlgorithmId, optional
@@ -631,6 +650,7 @@ cdef class OrderFactory:
         TimeInForce time_in_force = TimeInForce.GTC,
         datetime expire_time = None,
         bint reduce_only = False,
+        bint quote_quantity = False,
         TriggerType emulation_trigger = TriggerType.NO_TRIGGER,
         InstrumentId trigger_instrument_id = None,
         ExecAlgorithmId exec_algorithm_id = None,
@@ -658,6 +678,8 @@ cdef class OrderFactory:
             The order expiration (for ``GTD`` orders).
         reduce_only : bool, default False
             If the order carries the 'reduce-only' execution instruction.
+        quote_quantity : bool
+            If the order quantity is denominated in the quote currency.
         emulation_trigger : TriggerType, default ``NO_TRIGGER``
             The orders emulation trigger.
         trigger_instrument_id : InstrumentId, optional
@@ -700,6 +722,7 @@ cdef class OrderFactory:
             time_in_force=time_in_force,
             expire_time_ns=0 if expire_time is None else dt_to_unix_nanos(expire_time),
             reduce_only=reduce_only,
+            quote_quantity=quote_quantity,
             emulation_trigger=emulation_trigger,
             trigger_instrument_id=trigger_instrument_id,
             contingency_type=ContingencyType.NO_CONTINGENCY,
@@ -723,6 +746,7 @@ cdef class OrderFactory:
         datetime expire_time = None,
         bint post_only = False,
         bint reduce_only = False,
+        bint quote_quantity = False,
         Quantity display_qty = None,
         TriggerType emulation_trigger = TriggerType.NO_TRIGGER,
         InstrumentId trigger_instrument_id = None,
@@ -755,6 +779,8 @@ cdef class OrderFactory:
             If the order will only provide liquidity (make a market).
         reduce_only : bool, default False
             If the order carries the 'reduce-only' execution instruction.
+        quote_quantity : bool
+            If the order quantity is denominated in the quote currency.
         display_qty : Quantity, optional
             The quantity of the order to display on the public book (iceberg).
         emulation_trigger : TriggerType, default ``NO_TRIGGER``
@@ -803,6 +829,7 @@ cdef class OrderFactory:
             expire_time_ns=0 if expire_time is None else dt_to_unix_nanos(expire_time),
             post_only=post_only,
             reduce_only=reduce_only,
+            quote_quantity=quote_quantity,
             display_qty=display_qty,
             emulation_trigger=emulation_trigger,
             trigger_instrument_id=trigger_instrument_id,
@@ -827,6 +854,7 @@ cdef class OrderFactory:
         TimeInForce time_in_force = TimeInForce.GTC,
         datetime expire_time = None,
         bint reduce_only = False,
+        bint quote_quantity = False,
         TriggerType emulation_trigger = TriggerType.NO_TRIGGER,
         InstrumentId trigger_instrument_id = None,
         ExecAlgorithmId exec_algorithm_id = None,
@@ -859,6 +887,8 @@ cdef class OrderFactory:
             The order expiration (for ``GTD`` orders).
         reduce_only : bool, default False
             If the order carries the 'reduce-only' execution instruction.
+        quote_quantity : bool
+            If the order quantity is denominated in the quote currency.
         emulation_trigger : TriggerType, default ``NO_TRIGGER``
             The orders emulation trigger.
         exec_algorithm_id : ExecAlgorithmId, optional
@@ -903,6 +933,7 @@ cdef class OrderFactory:
             time_in_force=time_in_force,
             expire_time_ns=0 if expire_time is None else dt_to_unix_nanos(expire_time),
             reduce_only=reduce_only,
+            quote_quantity=quote_quantity,
             emulation_trigger=emulation_trigger,
             trigger_instrument_id=trigger_instrument_id,
             contingency_type=ContingencyType.NO_CONTINGENCY,
@@ -929,6 +960,7 @@ cdef class OrderFactory:
         datetime expire_time = None,
         bint post_only = False,
         bint reduce_only = False,
+        bint quote_quantity = False,
         Quantity display_qty = None,
         TriggerType emulation_trigger = TriggerType.NO_TRIGGER,
         InstrumentId trigger_instrument_id = None,
@@ -969,6 +1001,8 @@ cdef class OrderFactory:
             If the order will only provide liquidity (make a market).
         reduce_only : bool, default False
             If the order carries the 'reduce-only' execution instruction.
+        quote_quantity : bool
+            If the order quantity is denominated in the quote currency.
         display_qty : Quantity, optional
             The quantity of the order to display on the public book (iceberg).
         emulation_trigger : TriggerType, default ``NO_TRIGGER``
@@ -1022,6 +1056,7 @@ cdef class OrderFactory:
             expire_time_ns=0 if expire_time is None else dt_to_unix_nanos(expire_time),
             post_only=post_only,
             reduce_only=reduce_only,
+            quote_quantity=quote_quantity,
             display_qty=display_qty,
             emulation_trigger=emulation_trigger,
             trigger_instrument_id=trigger_instrument_id,
@@ -1050,6 +1085,7 @@ cdef class OrderFactory:
         datetime expire_time = None,
         bint entry_post_only = False,
         bint tp_post_only = True,
+        bint quote_quantity = False,
         TriggerType emulation_trigger = TriggerType.NO_TRIGGER,
         InstrumentId trigger_instrument_id = None,
         ContingencyType contingency_type = ContingencyType.OUO,
@@ -1097,6 +1133,8 @@ cdef class OrderFactory:
             If the entry order will only provide liquidity (make a market).
         tp_post_only : bool, default False
             If the take-profit order will only provide liquidity (make a market).
+        quote_quantity : bool
+            If the order quantity is denominated in the quote currency.
         emulation_trigger : TriggerType, default ``NO_TRIGGER``
             The emulation trigger type for the entry, as well as the TP and SL bracket orders.
         trigger_instrument_id : InstrumentId, optional
@@ -1140,6 +1178,7 @@ cdef class OrderFactory:
                 init_id=UUID4(),
                 ts_init=self._clock.timestamp_ns(),
                 time_in_force=TimeInForce.GTC,
+                quote_quantity=quote_quantity,
                 contingency_type=ContingencyType.OTO,
                 order_list_id=order_list_id,
                 linked_order_ids=[sl_client_order_id, tp_client_order_id],
@@ -1162,6 +1201,7 @@ cdef class OrderFactory:
                 time_in_force=time_in_force,
                 expire_time_ns=0 if expire_time is None else dt_to_unix_nanos(expire_time),
                 post_only=entry_post_only,
+                quote_quantity=quote_quantity,
                 emulation_trigger=emulation_trigger,
                 trigger_instrument_id=trigger_instrument_id,
                 contingency_type=ContingencyType.OTO,
@@ -1186,6 +1226,7 @@ cdef class OrderFactory:
                 ts_init=self._clock.timestamp_ns(),
                 time_in_force=time_in_force,
                 expire_time_ns=0 if expire_time is None else dt_to_unix_nanos(expire_time),
+                quote_quantity=quote_quantity,
                 emulation_trigger=emulation_trigger,
                 trigger_instrument_id=trigger_instrument_id,
                 contingency_type=ContingencyType.OTO,
@@ -1212,6 +1253,7 @@ cdef class OrderFactory:
                 time_in_force=time_in_force,
                 expire_time_ns=0 if expire_time is None else dt_to_unix_nanos(expire_time),
                 post_only=entry_post_only,
+                quote_quantity=quote_quantity,
                 emulation_trigger=emulation_trigger,
                 trigger_instrument_id=trigger_instrument_id,
                 contingency_type=ContingencyType.OTO,
@@ -1242,6 +1284,7 @@ cdef class OrderFactory:
                 time_in_force=TimeInForce.GTC,
                 post_only=tp_post_only,
                 reduce_only=True,
+                quote_quantity=quote_quantity,
                 display_qty=None,
                 emulation_trigger=emulation_trigger,
                 trigger_instrument_id=trigger_instrument_id,
@@ -1269,6 +1312,7 @@ cdef class OrderFactory:
                 time_in_force=TimeInForce.GTC,
                 post_only=tp_post_only,
                 reduce_only=True,
+                quote_quantity=quote_quantity,
                 display_qty=None,
                 emulation_trigger=emulation_trigger,
                 trigger_instrument_id=trigger_instrument_id,
@@ -1294,6 +1338,7 @@ cdef class OrderFactory:
                 ts_init=self._clock.timestamp_ns(),
                 time_in_force=TimeInForce.GTC,
                 reduce_only=True,
+                quote_quantity=quote_quantity,
                 emulation_trigger=emulation_trigger,
                 trigger_instrument_id=trigger_instrument_id,
                 contingency_type=contingency_type,
@@ -1323,6 +1368,7 @@ cdef class OrderFactory:
             ts_init=self._clock.timestamp_ns(),
             time_in_force=TimeInForce.GTC,
             reduce_only=True,
+            quote_quantity=quote_quantity,
             emulation_trigger=emulation_trigger,
             trigger_instrument_id=trigger_instrument_id,
             contingency_type=contingency_type,