+++ nautilus_trader/model/orders/stop_limit.pyx
@@ -94,6 +94,8 @@ cdef class StopLimitOrder(Order):
         If the ``LIMIT`` order will only provide liquidity (once triggered).
     reduce_only : bool, default False
         If the ``LIMIT`` order carries the 'reduce-only' execution instruction.
+    quote_quantity : bool, default False
+        If the order quantity is denominated in the quote currency.
     display_qty : Quantity, optional
         The quantity of the ``LIMIT`` order to display on the public book (iceberg).
     emulation_trigger : EmulationTrigger, default ``NO_TRIGGER``
@@ -155,6 +157,7 @@ cdef class StopLimitOrder(Order):
         uint64_t expire_time_ns = 0,
         bint post_only = False,
         bint reduce_only = False,
+        bint quote_quantity = False,
         Quantity display_qty = None,
         TriggerType emulation_trigger = TriggerType.NO_TRIGGER,
         InstrumentId trigger_instrument_id = None,
@@ -204,6 +207,7 @@ cdef class StopLimitOrder(Order):
             time_in_force=time_in_force,
             post_only=post_only,
             reduce_only=reduce_only,
+            quote_quantity=quote_quantity,
             options=options,
             emulation_trigger=emulation_trigger,
             trigger_instrument_id=trigger_instrument_id,
@@ -321,6 +325,7 @@ cdef class StopLimitOrder(Order):
             "status": self._fsm.state_string_c(),
             "is_post_only": self.is_post_only,
             "is_reduce_only": self.is_reduce_only,
+            "is_quote_quantity": self.is_quote_quantity,
             "display_qty": str(self.display_qty) if self.display_qty is not None else None,
             "emulation_trigger": trigger_type_to_str(self.emulation_trigger),
             "trigger_instrument_id": self.trigger_instrument_id.to_str() if self.trigger_instrument_id is not None else None,
@@ -377,6 +382,7 @@ cdef class StopLimitOrder(Order):
             ts_init=init.ts_init,
             post_only=init.post_only,
             reduce_only=init.reduce_only,
+            quote_quantity=init.quote_quantity,
             display_qty=Quantity.from_str_c(display_qty_str) if display_qty_str is not None else None,
             emulation_trigger=init.emulation_trigger,
             trigger_instrument_id=init.trigger_instrument_id,